Refactor deployment with proxies
