+++ cmd/elkia-auth-proxy/Dockerfile
