+++ cmd/elkia-auth/Dockerfile
