+++ manifests/elkia-auth-proxy/base/deployment.yaml
@@ -1,0 +1,36 @@
+apiVersion: apps/v1
+kind: Deployment
+metadata:
+  name: elkia-auth-proxy
+  namespace: elkia
+  labels:
+    app.kubernetes.io/name: elkia-auth-proxy
+    app.kubernetes.io/instance: elkia-auth-proxy
+spec:
+  replicas: 1
+  selector:
+    matchLabels:
+      app.kubernetes.io/name: elkia-auth-proxy
+      app.kubernetes.io/instance: elkia-auth-proxy
+  template:
+    metadata:
+      labels:
+        app.kubernetes.io/name: elkia-auth-proxy
+        app.kubernetes.io/instance: elkia-auth-proxy
+    spec:
+      containers:
+        - name: elkia-auth-proxy
+          image: ghcr.io/infinity-blackhole/elkia-auth-proxy
+          env:
+            - name: HOST
+              value: "0.0.0.0"
+          ports:
+            - containerPort: 4123
+              name: nslp
+          resources:
+            requests:
+              cpu: 10m
+              memory: 16Mi
+            limits:
+              cpu: 100m
+              memory: 128Mi
