+++ manifests/elkia-auth-proxy/base/svc.yaml
@@ -1,0 +1,18 @@
+apiVersion: v1
+kind: Service
+metadata:
+  name: elkia-auth-proxy
+  namespace: elkia
+  labels:
+    app.kubernetes.io/name: elkia-auth-proxy
+    app.kubernetes.io/instance: elkia-auth-proxy
+    fleet.elkia.io/port: nslp
+    fleet.elkia.io/component: auth
+spec:
+  ports:
+    - name: nslp
+      port: 4123
+      targetPort: nslp
+  selector:
+    app.kubernetes.io/name: elkia-auth-proxy
+    app.kubernetes.io/instance: elkia-auth-proxy
