+++ manifests/elkia-auth-proxy/components/elkia-fleet/kustomization.yaml
@@ -1,0 +1,4 @@
+apiVersion: kustomize.config.k8s.io/v1alpha1
+kind: Component
+patches:
+  - path: deployment.patch.yaml
