+++ manifests/elkia-auth-proxy/overlays/default/deployment.patch.yaml
@@ -1,0 +1,13 @@
+apiVersion: apps/v1
+kind: Deployment
+metadata:
+  name: elkia-auth-proxy
+  namespace: elkia
+spec:
+  template:
+    spec:
+      containers:
+        - name: elkia-auth-proxy
+          env:
+            - name: LOG_LEVEL
+              value: debug
