+++ manifests/elkia-auth-proxy/overlays/default/kustomization.yaml
@@ -1,0 +1,9 @@
+apiVersion: kustomize.config.k8s.io/v1beta1
+kind: Kustomization
+resources:
+  - ../../base
+components:
+  - ../../components/elkia-auth
+patches:
+  - path: deployment.patch.yaml
+  - path: svc.patch.yaml
