+++ manifests/elkia-auth-proxy/overlays/default/svc.patch.yaml
@@ -1,0 +1,7 @@
+apiVersion: v1
+kind: Service
+metadata:
+  name: elkia-auth-proxy
+  namespace: elkia
+spec:
+  type: LoadBalancer
