+++ manifests/elkia-auth/base/deployment.yaml
@@ -25,8 +25,8 @@ spec:
             - name: HOST
               value: "0.0.0.0"
           ports:
-            - containerPort: 4123
-              name: nslp
+            - containerPort: 8080
+              name: http
           resources:
             requests:
               cpu: 10m
