+++ manifests/elkia-auth/base/svc.yaml
@@ -6,13 +6,11 @@ metadata:
   labels:
     app.kubernetes.io/name: elkia-auth
     app.kubernetes.io/instance: elkia-auth
-    fleet.elkia.io/port: nslp
-    fleet.elkia.io/component: auth
 spec:
   ports:
-    - name: nslp
-      port: 4123
-      targetPort: nslp
+    - name: http
+      port: 80
+      targetPort: http
   selector:
     app.kubernetes.io/name: elkia-auth
     app.kubernetes.io/instance: elkia-auth
