+++ manifests/elkia-gateway-proxy/overlays/default/deployment.patch.yaml
@@ -1,0 +1,13 @@
+apiVersion: apps/v1
+kind: Deployment
+metadata:
+  name: elkia-gateway-proxy
+  namespace: elkia
+spec:
+  template:
+    spec:
+      containers:
+        - name: elkia-gateway-proxy
+          env:
+            - name: LOG_LEVEL
+              value: debug
