+++ manifests/elkia-gateway-proxy/overlays/default/svc.patch.yaml
@@ -1,0 +1,7 @@
+apiVersion: v1
+kind: Service
+metadata:
+  name: elkia-gateway-proxy
+  namespace: elkia
+spec:
+  type: LoadBalancer
