+++ skaffold.yaml
@@ -10,18 +10,30 @@ build:
       docker:
         target: build
         dockerfile: cmd/elkia-auth/Dockerfile
-    - image: ghcr.io/infinity-blackhole/elkia-fleet
+    - image: ghcr.io/infinity-blackhole-proxy/elkia-auth-proxy
       context: .
       runtimeType: go
       docker:
         target: build
-        dockerfile: cmd/elkia-fleet/Dockerfile
+        dockerfile: cmd/elkia-auth/Dockerfile
     - image: ghcr.io/infinity-blackhole/elkia-gateway
       context: .
       runtimeType: go
       docker:
         target: build
         dockerfile: cmd/elkia-gateway/Dockerfile
+    - image: ghcr.io/infinity-blackhole/elkia-gateway-proxy
+      context: .
+      runtimeType: go
+      docker:
+        target: build
+        dockerfile: cmd/elkia-gateway-proxy/Dockerfile
+    - image: ghcr.io/infinity-blackhole/elkia-fleet
+      context: .
+      runtimeType: go
+      docker:
+        target: build
+        dockerfile: cmd/elkia-fleet/Dockerfile
   local:
     useBuildkit: true
 manifests:
